Change VisualStudioDescriptionFactory to interface
