+++ src/Razor/test/Microsoft.VisualStudio.Editor.Razor.Test/Completion/DefaultVisualStudioDescriptionFactoryTest.cs
@@ -23,7 +23,7 @@ public class DefaultVisualStudioDescriptionFactoryTest : TestBase
     public void CreateClassifiedDescription_SingleDescription_NoSeparator()
     {
         // Arrange
-        var factory = new DefaultVisualStudioDescriptionFactory();
+        var factory = new VisualStudioDescriptionFactory();
         var description = new AggregateBoundAttributeDescription(new[]
         {
             new BoundAttributeDescriptionInfo("TheReturnType", "TheTypeName", "ThePropertyName", "The documentation"),
@@ -33,14 +33,14 @@ public class DefaultVisualStudioDescriptionFactoryTest : TestBase
         var result = factory.CreateClassifiedDescription(description);
 
         // Assert
-        Assert.DoesNotContain(DefaultVisualStudioDescriptionFactory.SeparatorElement, result.Elements);
+        Assert.DoesNotContain(VisualStudioDescriptionFactory.SeparatorElement, result.Elements);
     }
 
     [Fact]
     public void CreateClassifiedDescription_MultipleDescription_Separator()
     {
         // Arrange
-        var factory = new DefaultVisualStudioDescriptionFactory();
+        var factory = new VisualStudioDescriptionFactory();
         var description = new AggregateBoundAttributeDescription(new[]
         {
             new BoundAttributeDescriptionInfo("TheReturnType", "TheTypeName", "ThePropertyName", "The documentation"),
@@ -51,14 +51,14 @@ public class DefaultVisualStudioDescriptionFactoryTest : TestBase
         var result = factory.CreateClassifiedDescription(description);
 
         // Assert
-        Assert.Contains(DefaultVisualStudioDescriptionFactory.SeparatorElement, result.Elements);
+        Assert.Contains(VisualStudioDescriptionFactory.SeparatorElement, result.Elements);
     }
 
     [Fact]
     public void CreateClassifiedDescription_RepresentsReturnType()
     {
         // Arrange
-        var factory = new DefaultVisualStudioDescriptionFactory();
+        var factory = new VisualStudioDescriptionFactory();
         var description = new AggregateBoundAttributeDescription(new[]
         {
             new BoundAttributeDescriptionInfo("TheReturnType", "TheTypeName", "ThePropertyName", "The documentation"),
@@ -76,7 +76,7 @@ public class DefaultVisualStudioDescriptionFactoryTest : TestBase
     public void CreateClassifiedDescription_RepresentsTypeName()
     {
         // Arrange
-        var factory = new DefaultVisualStudioDescriptionFactory();
+        var factory = new VisualStudioDescriptionFactory();
         var description = new AggregateBoundAttributeDescription(new[]
         {
             new BoundAttributeDescriptionInfo("TheReturnType", "TheTypeName", "ThePropertyName", "The documentation"),
@@ -94,7 +94,7 @@ public class DefaultVisualStudioDescriptionFactoryTest : TestBase
     public void CreateClassifiedDescription_RepresentsPropertyName()
     {
         // Arrange
-        var factory = new DefaultVisualStudioDescriptionFactory();
+        var factory = new VisualStudioDescriptionFactory();
         var description = new AggregateBoundAttributeDescription(new[]
         {
             new BoundAttributeDescriptionInfo("TheReturnType", "TheTypeName", "ThePropertyName", "The documentation"),
@@ -112,7 +112,7 @@ public class DefaultVisualStudioDescriptionFactoryTest : TestBase
     public void CreateClassifiedDescription_RepresentsDocumentation()
     {
         // Arrange
-        var factory = new DefaultVisualStudioDescriptionFactory();
+        var factory = new VisualStudioDescriptionFactory();
         var description = new AggregateBoundAttributeDescription(new[]
         {
             new BoundAttributeDescriptionInfo("TheReturnType", "TheTypeName", "ThePropertyName", "The documentation"),
@@ -130,7 +130,7 @@ public class DefaultVisualStudioDescriptionFactoryTest : TestBase
     public void CreateClassifiedDescription_CanSimplifyKeywordReturnTypes()
     {
         // Arrange
-        var factory = new DefaultVisualStudioDescriptionFactory();
+        var factory = new VisualStudioDescriptionFactory();
         var description = new AggregateBoundAttributeDescription(new[]
         {
             new BoundAttributeDescriptionInfo("System.String", "TheTypeName", "ThePropertyName", "The documentation"),
@@ -149,7 +149,7 @@ public class DefaultVisualStudioDescriptionFactoryTest : TestBase
     public void CreateClassifiedDescription_CanRepresentMultipleDescriptions()
     {
         // Arrange
-        var factory = new DefaultVisualStudioDescriptionFactory();
+        var factory = new VisualStudioDescriptionFactory();
         var description = new AggregateBoundAttributeDescription(new[]
         {
             new BoundAttributeDescriptionInfo("System.String", "TheTypeName", "ThePropertyName", "The documentation"),
