+++ src/Razor/test/Microsoft.VisualStudio.Editor.Razor.Test/Completion/RazorDirectiveAttributeCompletionSourceTest.cs
@@ -46,7 +46,7 @@ public class RazorDirectiveAttributeCompletionSourceTest : ProjectSnapshotManage
         // Arrange
         var expectedResult = new ContainerElement(ContainerElementStyle.Wrapped);
         var description = new AggregateBoundAttributeDescription(Array.Empty<BoundAttributeDescriptionInfo>());
-        var descriptionFactory = Mock.Of<VisualStudioDescriptionFactory>(factory => factory.CreateClassifiedDescription(description) == expectedResult, MockBehavior.Strict);
+        var descriptionFactory = Mock.Of<IVisualStudioDescriptionFactory>(factory => factory.CreateClassifiedDescription(description) == expectedResult, MockBehavior.Strict);
         var source = new RazorDirectiveAttributeCompletionSource(
             Dispatcher,
             Mock.Of<VisualStudioRazorParser>(MockBehavior.Strict),
@@ -242,7 +242,7 @@ public class RazorDirectiveAttributeCompletionSourceTest : ProjectSnapshotManage
             Mock.Of<VisualStudioRazorParser>(MockBehavior.Strict),
             Mock.Of<IRazorCompletionFactsService>(MockBehavior.Strict),
             Mock.Of<ICompletionBroker>(MockBehavior.Strict),
-            Mock.Of<VisualStudioDescriptionFactory>(MockBehavior.Strict),
+            Mock.Of<IVisualStudioDescriptionFactory>(MockBehavior.Strict),
             JoinableTaskFactory);
         return source;
     }
